functions/requirements.txt
azure-functions
azure-identity
azure-cosmos
azure-storage-queue
requests